api/requirements.txt
argon2-cffi==21.3.0
cryptocode==0.1
daphne==3.0.2
Django==3.2.9
django-filter==21.1
django-ipware==4.0.2
django-phonenumber-field[phonenumberslite]==6.0.0
django-redis==5.2.0
django-rest-knox==4.1.0
django-user-agents==0.4.0
djangorestframework==3.13.1
freezegun==1.1.0
psycopg2-binary==2.9.3
python-decouple==3.5
twilio==7.5.0
